Renamed and moved conversion functions to utils
