+++ hummingbot/connector/exchange/bloxroute_openbook/bloxroute_openbook_utils.py
@@ -5,10 +5,12 @@ from typing import Any, Callable, Coroutine
 import betterproto
 import bxsolana_trader_proto.api as api
 import bxsolana_trader_proto.common as common
+from bxsolana_trader_proto import api
 from pydantic import Field, SecretStr
 
 from hummingbot.client.config.config_data_types import BaseConnectorConfigMap, ClientFieldData
 from hummingbot.core.data_type.common import OrderType, TradeType
+from hummingbot.core.data_type.in_flight_order import OrderState
 from hummingbot.core.data_type.trade_fee import TradeFeeSchema
 
 EXAMPLE_PAIR = "SOL-USDC"
@@ -49,7 +51,7 @@ class BloXrouteConnectorMap(BaseConnectorConfigMap):
 KEYS = BloXrouteConnectorMap.construct()
 
 
-def trade_type_to_side(trade_type: TradeType) -> api.Side:
+def convert_hbot_trade_type(trade_type: TradeType) -> api.Side:
     if trade_type == TradeType.BUY:
         return api.Side.S_BID
     elif trade_type == TradeType.SELL:
@@ -58,7 +60,7 @@ def trade_type_to_side(trade_type: TradeType) -> api.Side:
         return api.Side.S_UNKNOWN
 
 
-def order_type_to_blxr_order_type(order_type: OrderType) -> common.OrderType:
+def convert_hbot_order_type(order_type: OrderType) -> common.OrderType:
     if order_type == OrderType.MARKET:
         return common.OrderType.OT_MARKET
     elif order_type == OrderType.LIMIT:
@@ -66,11 +68,35 @@ def order_type_to_blxr_order_type(order_type: OrderType) -> common.OrderType:
     else:
         raise Exception(f"unknown order type {order_type.value}")
 
+
+def convert_hbot_client_order_id(client_order_id: str):
+    num = _convert_to_number(client_order_id)
+    return truncate(num, 7)
+
+
+def _convert_to_number(s):
+    return int.from_bytes(s.encode(), "little")
+
+
+def convert_blxr_order_status(order_status: api.OrderStatus) -> OrderState:
+    if order_status == api.OrderStatus.OS_OPEN:
+        return OrderState.OPEN
+    elif order_status == api.OrderStatus.OS_PARTIAL_FILL:
+        return OrderState.PARTIALLY_FILLED
+    elif order_status == api.OrderStatus.OS_FILLED:
+        return OrderState.FILLED
+    elif order_status == api.OrderStatus.OS_CANCELLED:
+        return OrderState.CANCELED
+    else:
+        return OrderState.PENDING_CREATE
+
+
 # gets the last n digits of a number
 def truncate(num: int, n: int) -> int:
     num_str = str(num)
     trunc_num_str = num_str[-n:]
     return int(trunc_num_str)
+
 
 T = typing.TypeVar("T")
